motd broadcast added and renamed to editmotd.
